Bug 1839762 - Cast scoped enums to their underlying type when printing them. r=media-playback-reviewers,necko-reviewers,application-update-reviewers,geckoview-reviewers,handyman,emilio,valentin,bytesized,owlish,karlt
There is no implicit conversion for scoped enums, so using them without
an explicit conversion in varargs functions is undefined behavior. GCC
has had a warning about this for a long while, but clang only gained
this a few days ago on trunk.

Differential Revision: https://phabricator.services.mozilla.com/D181723

UltraBlame original commit: 7898f66c9d4cdd0ab2298cfe0d0f8b0e196c73b8

+++ netwerk/system/win32/nsNotifyAddrListener.cpp
@@ -26,7 +26,6 @@
 #include <netioapi.h>
 #include <netlistmgr.h>
 #include <iprtrmib.h>
-#include "plstr.h"
 #include "mozilla/Logging.h"
 #include "nsComponentManagerUtils.h"
 #include "nsThreadUtils.h"
@@ -259,7 +258,8 @@ void nsNotifyAddrListener::calculateNetworkId(void) {
       mNetworkId.Truncate();
     }
     Telemetry::Accumulate(Telemetry::NETWORK_ID2, 0);
-    LOG(("calculateNetworkId: no network ID Base64Encode error %X", rv));
+    LOG(("calculateNetworkId: no network ID Base64Encode error %X",
+         uint32_t(rv)));
     return;
   }
